requirements.txt
fastapi
uvicorn
pytesseract
opencv-python-headless
numpy
Pillow
requests
pillow-heif